deploy headers token for Api questions

+++ backend/Controllers/questionController.js
@@ -1,18 +1,34 @@
 const questionModel=require('../Models/questionModel');
-
+require('dotenv').config();
+const jwt=require('jsonwebtoken')
 module.exports.allQuestions=async(req,res,next)=>{
     res.setHeader("Access-Control-Allow-Origin", "http://localhost:3000")
     res.setHeader("Access-Control-Allow-Credentials", "true");
     res.setHeader("Access-Control-Max-Age", "1800");
     res.setHeader("Access-Control-Allow-Headers", "content-type");
     res.setHeader( "Access-Control-Allow-Methods", "PUT, POST, GET, DELETE, PATCH, OPTIONS" );
-    await questionModel.find({}).then((data)=>{
-        res.send(data);
-        console.log(data)
-    }).catch((e)=>{
-        console.log(e)
+    let token=req.header("x-auth-token");
+    if(!token) return res.status(401).send({
+        ok:false,
+        error:"access denied . No token Provided"
     })
-    next();
+    try {
+        let decodedtoken=jwt.verify(token,process.env.privatekey);
+        req.user=decodedtoken;
+        await questionModel.find({}).then((data)=>{
+            res.send(data);
+            console.log(data)
+        }).catch((e)=>{
+            console.log(e)
+        })
+        next();
+    } catch (error) {
+        return res.status(401).send({
+            ok:false,
+            error:"access denied . No token Provided"
+        })
+    }
+    
 }
 module.exports.postQuestion=async(req,res,next)=>{
     res.setHeader("Access-Control-Allow-Origin", "http://localhost:3000")
@@ -25,12 +41,28 @@ module.exports.postQuestion=async(req,res,next)=>{
     if(!req.body){
         res.send({body:null})
     }
+
     else{
-        await questionModel.create(req.body).then((data)=>{
-            res.send(data)
-        }).then((e)=>{
-            console.log(e)
+        let token=req.header("x-auth-token");
+        if(!token) return res.status(401).send({
+            ok:false,
+            error:"access denied . No token Provided"
         })
+        try {
+            let decodedtoken=jwt.verify(token,process.env.privatekey);
+            req.user=decodedtoken;
+            await questionModel.create(req.body).then((data)=>{
+                res.send(data)
+            }).then((e)=>{
+                console.log(e)
+            })
+        } catch (error) {
+            res.status(401).send({
+                ok:false,
+                error:"access denied . No token Provided"
+            })
+        }
+        
     
     }
 }
@@ -41,12 +73,31 @@ module.exports.modifyQuestion=(req,res,next)=>{
     res.setHeader("Access-Control-Max-Age", "1800");
     res.setHeader("Access-Control-Allow-Headers", "content-type");
     res.setHeader( "Access-Control-Allow-Methods", "PUT, POST, GET, DELETE, PATCH, OPTIONS" );
-    questionModel.findByIdAndUpdate({_id:req.params.id},req.body).then((data)=>{
-        res.send(data);
-    }).catch((e)=>{
-        console.log(e);
-    });
-    next();
+    if(!token) return res.status(401).send({
+        ok:false,
+        error:"access denied . No token Provided"
+    })
+    else{
+        try {
+            let decodedtoken=jwt.verify(token,process.env.privatekey);
+            req.user=decodedtoken;
+            questionModel.findByIdAndUpdate({_id:req.params.id},req.body).then((data)=>{
+                res.send(data);
+            }).catch((e)=>{
+                console.log(e);
+            });
+            next();
+
+        } catch (error) {
+            return res.status(401).send({
+                ok:false,
+                error:"access denied . No token Provided"
+            })
+        }
+
+        
+    }
+    
 }
 
 module.exports.deleteQuestion=(req,res,next)=>{
@@ -55,9 +106,26 @@ module.exports.deleteQuestion=(req,res,next)=>{
     res.setHeader("Access-Control-Max-Age", "1800");
     res.setHeader("Access-Control-Allow-Headers", "content-type");
     res.setHeader( "Access-Control-Allow-Methods", "PUT, POST, GET, DELETE, PATCH, OPTIONS" );
-    questionModel.findByIdAndDelete(req.params.id).then(()=>{
-        console.log(req.params.id+" is deleted succesfully");
-    }).catch((e)=>{
-        console.log(e);
+    if(!token) return res.status(401).send({
+        ok:false,
+        error:"access denied . No token Provided"
     })
+    else{
+        
+        try {
+            let decodedtoken=jwt.verify(token,process.env.privatekey);
+            req.user=decodedtoken;
+            questionModel.findByIdAndDelete(req.params.id).then(()=>{
+                console.log(req.params.id+" is deleted succesfully");
+            }).catch((e)=>{
+                console.log(e);
+            })
+        } catch (error) {
+            return res.status(401).send({
+                ok:false,
+                error:"access denied . No token Provided"
+            })
+        }
+    }
+    
 }
